5. Framework/04_SpringMVC/src/test/java/com/kh/mvc/member/controller/MemberControllerTest.java
package com.kh.mvc.member.controller;

import static org.hamcrest.CoreMatchers.is;
import static org.springframework.test.web.client.match.MockRestRequestMatchers.content;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

@ExtendWith(SpringExtension.class)
@ContextConfiguration(
	locations = {
			"file:src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml",
			"file:src/main/webapp/WEB-INF/spring/root-context.xml"
	}
)
class MemberControllerTest {
	@Autowired
	MemberController controller;
	
	MockMvc mockMvc;
	
	@BeforeEach
	public void setUp() {
		mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
	}

	@Test
	public void enrollTest() throws Exception {		
		mockMvc.perform(get("/member/enroll"))
			   .andExpect(view().name("member/enroll"))
			   .andExpect(status().isOk());
	}
	
	@Test
	public void idCheckTest() throws Exception {
		mockMvc.perform(post("/member/idCheck").param("userId", "ismoon"))
			   .andExpect(status().isOk())
			   .andExpect(jsonPath("$.duplicate", is(true)));
	}
}
